VMS-8038: Added proxy timers for RADASS testing with default Qt-based implementations
<==== Mercurial data ====>
Branch: x_VMS-8038_create_RADASS_tests

+++ vms/client/nx_vms_client_desktop/src/camera/cam_display.cpp
@@ -1281,7 +1281,7 @@ void QnCamDisplay::processMetadata(const QnAbstractCompressedMetadataPtr& metada
 {
     if (metadata->metadataType == MetadataType::MediaStreamEvent)
     {
-        QByteArray data = QByteArray::fromRawData(metadata->data(), metadata->dataSize());
+        QByteArray data = QByteArray::fromRawData(metadata->data(), (int) metadata->dataSize());
         auto mediaEvent = QnLexical::deserialized<Qn::MediaStreamEvent>(
             QString::fromLatin1(data));
